Fixed pathfinding bug
When an enemy is destroying, during the death state (which has not bee fully implemented yet) the unit was still picking it up at attackable. As a result was being given two paths, may need to consider this for when an enemy if destroyed but there is an enemy next to it, may need to implement additional logic.